BBC News Web Scraping and Sentiment Analysis Model
A project I built in Python that web scrapes the news site BBC News and prints out the current headlines. It then saves the output to a .txt file. From there, I use a sentiment analysis model in order to determine if the headlines are generally positive, negative, or neutral. The compound score is a normalized score that ranges from -1 (most negative) to 1 (most positive), where a score of 0 indicates a neutral sentiment.